Set secure cookie flag properly with WebAuthSSLReturn
If WebAuthSSLReturn is set to true, we may see non-SSL connections
that are SSL from the perspective of the browser (such as a WebAuth
server behind an L7 load balancer that does SSL termination).  In
this case, we still want to set the secure flag on the cookie so that
the browser properly restricts it to SSL connections.

Trigger setting the secure flag off the combination of whether the
request is SSL and whether WebAuthSSLReturn is set, rather than
just the former.

Gbp-Pq: Name 0004-Set-secure-cookie-flag-properly-with-WebAuthSSLRetur.patch